+++ src/pages/api/faucet.ts
@@ -29,25 +29,25 @@ export default async function handler(req, res) {
     (req.headers['x-forwarded-for'] || '').split(',').pop().trim() || req.socket.remoteAddress
 
   if (!address) {
-    return res.status(400).json({ message: 'addressIsRequired' })
+    return res.status(400).json({ message: 'getFaucetErrorMessages.addressIsRequired' })
   }
 
   if (!passcode) {
-    return res.status(400).json({ message: 'passcodeIsRequired' })
+    return res.status(400).json({ message: 'getFaucetErrorMessages.passcodeIsRequired' })
   }
 
   if (!ethers.isAddress(address)) {
-    return res.status(400).json({ message: 'invalidAddress' })
+    return res.status(400).json({ message: 'getFaucetErrorMessages.invalidAddress' })
   }
 
   if (creatorPasscodes.length === 0) {
-    return res.status(500).json({ message: 'noPasscodes' })
+    return res.status(500).json({ message: 'getFaucetErrorMessages.noPasscodes' })
   }
 
   const foundPasscode = creatorPasscodes.find(creator => creator.passcode === passcode)
 
   if (foundPasscode === undefined) {
-    return res.status(400).json({ message: 'invalidPasscode' })
+    return res.status(400).json({ message: 'getFaucetErrorMessages.invalidPasscode' })
   }
 
   const canSendEth =
@@ -57,13 +57,13 @@ export default async function handler(req, res) {
   const faucetWalletBalance = await chain.provider.getBalance(faucetWallet.address)
 
   if (!canSendEth || faucetWalletBalance === 0n) {
-    return res.status(400).json({ message: 'faucetIsEmpty' })
+    return res.status(400).json({ message: 'getFaucetErrorMessages.faucetIsEmpty' })
   }
 
   const hasAlreadySendFaucet = foundPasscode.accountsDistributed.includes(address)
 
   if (hasAlreadySendFaucet) {
-    return res.status(400).json({ message: 'addressHasAlready' })
+    return res.status(400).json({ message: 'getFaucetErrorMessages.addressHasAlready' })
   }
 
   const transaction = await faucetWallet.sendTransaction({
